% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-application.R
\name{export_result}
\alias{export_result}
\title{Renderiza um AnalysisResult no formato pedido}
\usage{
export_result(result, format = c("plotly", "json"))
}
\arguments{
\item{format}{"plotly" ou "json"}
}
\description{
Renderiza um AnalysisResult no formato pedido
}
