% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-semantic-school-indicator-model.R
\name{indicator_label}
\alias{indicator_label}
\title{Rótulo humano de um eixo/dimensão do modelo, com fallback sensato.}
\usage{
indicator_label(model, which = c("value", "x", "y", "group"))
}
\arguments{
\item{which}{um de "value", "x", "y", "group"}
}
\description{
Rótulo humano de um eixo/dimensão do modelo, com fallback sensato.
}
