% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-application.R
\name{load_dataset}
\alias{load_dataset}
\title{Carrega um SchoolIndicatorModel a partir de qualquer DataSource
(memory_source, postgres_source, ou uma futura parquet_source/csv_source)}
\usage{
load_dataset(source, ...)
}
\description{
Carrega um SchoolIndicatorModel a partir de qualquer DataSource
(memory_source, postgres_source, ou uma futura parquet_source/csv_source)
}
