% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-datasource-postgres-source.R
\name{load_school_indicator.postgres_source}
\alias{load_school_indicator.postgres_source}
\title{Carrega os valores de UM indicador direto do Postgres}
\usage{
\method{load_school_indicator}{postgres_source}(
  source,
  indicator_id,
  rede = "todas",
  municipio_id = NULL,
  ...
)
}
\arguments{
\item{source}{objeto \code{postgres_source}}

\item{indicator_id}{um dos valores em \code{INDICATOR_WHITELIST}}

\item{rede}{um dos valores em \code{NETWORK_WHITELIST}}

\item{municipio_id}{inteiro ou \code{NULL}}
}
\description{
Carrega os valores de UM indicador direto do Postgres
}
