% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-datasource-memory-source.R
\name{load_school_indicator}
\alias{load_school_indicator}
\title{Generic: carrega um SchoolIndicatorModel a partir de uma fonte}
\usage{
load_school_indicator(source, ...)
}
\description{
Generic: carrega um SchoolIndicatorModel a partir de uma fonte
}
