% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-semantic-school-indicator-model.R
\name{new_school_indicator_model}
\alias{new_school_indicator_model}
\title{Constrói um SchoolIndicatorModel}
\usage{
new_school_indicator_model(data, metadata = list())
}
\arguments{
\item{data}{data.frame com colunas entre: \code{value}, \code{x_value}, \code{y_value}, \code{group}}

\item{metadata}{lista com rótulos: \code{value_label}, \code{x_label}, \code{y_label}, \code{group_label}}
}
\description{
Constrói um SchoolIndicatorModel
}
