% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repository-postgres.R
\name{persist}
\alias{persist}
\title{Persistir resultado de análise}
\usage{
persist(repository, result, ...)
}
\arguments{
\item{repository}{Objeto de repositório.}

\item{result}{Objeto \code{analysis_result} a ser gravado.}

\item{...}{Parâmetros adicionais específicos do repositório.}
}
\value{
O objeto \code{result} invisivelmente.
}
\description{
Método genérico S3 para gravar um \code{analysis_result} em um repositório.
}
