% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-application.R
\name{persist_result}
\alias{persist_result}
\title{Persiste um AnalysisResult usando um Repository}
\usage{
persist_result(result, repository, ...)
}
\description{
Persiste um AnalysisResult usando um Repository
}
