% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repository-postgres.R
\name{postgres_repository}
\alias{postgres_repository}
\title{Criar repositório PostgreSQL}
\usage{
postgres_repository(con)
}
\arguments{
\item{con}{Conexão ativa do DBI (PostgreSQL).}
}
\value{
Objeto de classe \code{postgres_repository}.
}
\description{
Instancia um objeto S3 \code{postgres_repository} para persistência de dados.
}
