% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-datasource-postgres-source.R
\name{postgres_source}
\alias{postgres_source}
\title{Cria uma fonte de dados Postgres (uso batch/offline)}
\usage{
postgres_source(con)
}
\arguments{
\item{con}{conexão DBI ativa (ex.: \code{RPostgres::dbConnect(...)})}
}
\description{
Cria uma fonte de dados Postgres (uso batch/offline)
}
