% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view-plotly.R
\name{render_plotly}
\alias{render_plotly}
\title{Renderiza um AnalysisResult como especificação Plotly (data + layout)}
\usage{
render_plotly(result)
}
\description{
Renderiza um AnalysisResult como especificação Plotly (data + layout)
}
