% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\name{stop_invalid_dataset}
\alias{stop_invalid_dataset}
\title{Modelo semântico carregado, mas sem os campos que a análise exige}
\usage{
stop_invalid_dataset(message)
}
\description{
Modelo semântico carregado, mas sem os campos que a análise exige
}
