% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\name{stop_invalid_parameter}
\alias{stop_invalid_parameter}
\title{Parâmetro de chamada inválido (ex.: \code{model} não é um modelo semântico)}
\usage{
stop_invalid_parameter(message)
}
\description{
Parâmetro de chamada inválido (ex.: \code{model} não é um modelo semântico)
}
